search works. Able to isolate user choice. Still need USER link
