+++ mysite/movies/urls.py
@@ -4,6 +4,6 @@ from . import views
 urlpatterns = [
     #
     path('movie_search', views.movie_search, name='movie-search'),
-    #path('display/', views.display, name='display')
-    path('movie_display', views.movie_display, name='movie-display')
+    path('movie_display', views.movie_display, name='movie-display'),
+    path('save_movie', views.save_movie, name='save-movie')
 ]
