+++ mysite/templates/index.html
@@ -1,12 +1,8 @@
 {% extends "base_generic.html" %}
 
 {% block content %}
-  <h1>Your Movie AI  Home</h1>
-  <p>Welcome to Movie AI, a website developed by <em>Team 1</em>!</p>
-  <h2>We decide your taste!</h2>
-  <p>The decider has the following record counts:</p>
-  <ul>
-    <li><strong>Date:</strong> {{ today }}</li>
-    <li><strong>Copies:</strong> {{ num_instances }}</li>
-  </ul>
+  <h1>Your Movie Home</h1>
+  <p>Welcome to Movie, a website developed by <em>Team 1</em>!</p>
+  <h2>We store your tastes, like a refrigerator for movies!</h2>
+  <p>Today is {{today}} </p>
 {% endblock %}
