+++ mysite/users/urls.py
@@ -6,5 +6,4 @@ urlpatterns = [
     #
     path('', views.home, name='home'),
     path('sign-up/', views.sign_up, name='sign-up'),
-    path('settings/', views.settings, name='settings')
 ]
